added breeze package , authentication , secured routes

+++ laravel11-crash-course-the_codeholic/app/Http/Controllers/NoteController.php
@@ -12,9 +12,9 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy('created_at','desc')->paginate();
+        $notes = Note::query()->where('user_id', request()->user()->id)->orderBy('created_at', 'desc')->paginate();
         // dd($notes);
-        return view('note.index',['notes'=>$notes]);
+        return view('note.index', ['notes' => $notes]);
     }
 
     /**
@@ -30,12 +30,12 @@ class NoteController extends Controller
      */
     public function store(Request $request)
     {
-        $data=$request->validate([
-           'note'=>['required','string']
+        $data = $request->validate([
+            'note' => ['required', 'string']
         ]);
-        $data['user_id']=1;
+        $data['user_id'] = $request->user()->id;
 
-        $note=Note::create($data);
+        $note = Note::create($data);
 
         return to_route('note.show', $note->id)->with('mess ', 'Note created successfully');
     }
@@ -45,8 +45,10 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
-      
-        return view('note.show',['note'=>$note]);
+        if ($note->user_id != request()->user()->id) {
+            abort(403);
+        }
+        return view('note.show', ['note' => $note]);
     }
 
     /**
@@ -54,7 +56,7 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
-        return view('note.edit',['note'=>$note]);
+        return view('note.edit', ['note' => $note]);
     }
 
     /**
@@ -62,10 +64,13 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
-       $data=$request->validate([
-           'note'=>['required','string']
+        if ($note->user_id != request()->user()->id) {
+            abort(403);
+        }
+        $data = $request->validate([
+            'note' => ['required', 'string']
         ]);
-        
+
         $note->update($data);
 
         return to_route('note.show', $note->id)->with('message', 'Note updated successfully');
@@ -76,6 +81,9 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id != request()->user()->id) {
+            abort(403);
+        }
         $note->delete();
 
         return to_route('note.index')->with('message', 'Note deleted successfully');
